% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_erb.R
\name{add_erb}
\alias{add_erb}
\title{Increase Hz value by erbs}
\usage{
add_erb(hz_vals, erbs)
}
\arguments{
\item{hz_vals}{Numeric vector of Hz values}

\item{erbs}{Numeric vector of ERB values to add to the Hz values}
}
\value{
Matrix where columns correspond to input \code{hz_vals} indices and rows
correspond to \code{erbs} indices
}
\description{
Given a vector of Hz values, return a matrix obtained by adding to the values
in specified ERB increments.
}
\examples{

add_erb(c(70, 110), 1:3)
}
