% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tier.R
\name{add_tier}
\alias{add_tier}
\title{Add tier to existing textgrid}
\usage{
add_tier(textgrid, new_tier_list, filename = NULL, ...)
}
\arguments{
\item{textgrid}{Textgrid object to add to}

\item{new_tier_list}{New tiers to add, durations must be contained within
the tmin and tmax of the textgrid}

\item{filename}{New filename for the textgrid. If \code{NULL}, the default,
the filename will be the same as that of the input \code{textgrid}}

\item{...}{Additional logical arguments passed to \code{as_textgrid}}
}
\value{
Textgrid object with new tiers added at the bottom.
}
\description{
Given a textgrid object (like one returned by \code{rPraat::tg.read}) add new tiers
speciied by \code{new_tier_list}. The new tiers will be added to the bottom of the
textgrid. This function is useful if you calculate new interval or point
tiers programmatically with R and want to add them back to a textgrid.
See \code{rPraat::tg.write} for an existing implementation to write the new
textgrid and \code{sosprosody::write_textgrid} for a wrapper that fills the
filename parameter from the textgrid object directly.
}
