% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_textgrid.R
\name{as_textgrid}
\alias{as_textgrid}
\title{Convert list to textgrid object}
\usage{
as_textgrid(
  tier_list,
  filename = "",
  tmin = 0,
  tmax = NULL,
  .strict_t1 = TRUE,
  .strict_t2 = TRUE,
  .quiet = FALSE
)
}
\arguments{
\item{tier_list}{Named list with relevant tier information, see \code{is_tier_list}}

\item{filename}{String, filename to use, defaults to \code{""}}

\item{tmin}{Numeric, time min for the textgrid, defaults to \code{0}}

\item{tmax}{Numeric, time max for the textgrid, if \code{NULL} (the default),
this value will be extracted from an interval tier.}

\item{.strict_t1}{Logical, default \code{TRUE} whether to ensure that every
interval tier starts at \code{tmin}. May give unexpected results or corrupted
textgrids if set to \code{FALSE}. Will also check that the new \code{tmin} is <= the
first time point in each tier (t1 for intervals).}

\item{.strict_t2}{Logical, default \code{TRUE} whether to ensure that every
interval tier ends at \code{tmax}. May give unexpected results or corrupted
textgrids if set to \code{FALSE}. Will also check that the new \code{tmax} is >= the
last time point in each tier (t2 for intervals).}

\item{.quiet}{Whether to warn user if  \code{t1} or \code{t2} values are changed as a result of
\code{.strict_t1} and \code{.strict_t2}}
}
\value{
TextGrid object with the given filename
}
\description{
Similar to \code{rPraat::as.tg} but with more validation on the input list.
The output of this can be passed to \code{add_tier} with another textgrid.
}
\details{
Note that the data types of each part of the tier will be coerced to the
correct value automatically via \code{coerce_tier_types}. Unexpected results may
occur if you set up a tier list with incorrect or missing values.
}
