% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_pitchtiers.R
\name{batch_process_pitchtiers}
\alias{batch_process_pitchtiers}
\title{Load multiple pitchtiers into wrangled dataframe}
\usage{
batch_process_pitchtiers(ptdir, parallelize = FALSE)
}
\arguments{
\item{ptdir}{Directory containing pitchtier files}

\item{parallelize}{Logical, whether to run in parallel with \code{furrr::future_map_dfr}}
}
\value{
Dataframe of the pitchtier information from each file
}
\description{
Wrapper around filepath |> pitchtier_to_dataframe()
merged together via \code{purrr::map_dfr}
}
