% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_textgrid.R
\name{coerce_tier_types}
\alias{coerce_tier_types}
\title{Coerce tier data types to correct types}
\usage{
coerce_tier_types(tier_list)
}
\arguments{
\item{tier_list}{List representation of a textgrid tier, see \code{is_tier_list}}
}
\value{
The input tier list but coerced to the correct datatypes
}
\description{
Helper for \code{as_textgrid}, will ensure that the data types are correct
for textgrids. You should use this on a tier list before passing to
\code{is_tier_list}.
}
