% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groupings.R
\name{.get_groupings}
\alias{.get_groupings}
\title{Get groupings from user and dataframe}
\usage{
.get_groupings(grpd_df, .grouping, .overridegroups = FALSE)
}
\arguments{
\item{grpd_df}{Dataframe, perhaps with groups}

\item{.grouping}{Character vector}

\item{.overridegroups}{Logical, defaults to \code{FALSE} to merge the existing
grouping structure and the user's specification. Set to \code{TRUE} to drop the
existing groups and group by the user specification.}
}
\value{
Character vector with no duplicates, either \code{.grouping} or the union
of \code{.grouping} and the groups of \code{grpd_df}.
}
\description{
Helper function, extracts the groups from a grouped dataframe and the
user supplied groupings, then either merges them or returns the user's
preferences.
}
