% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{dt.write}
\alias{dt.write}
\title{dt.write}
\usage{
dt.write(pt, fileNameDurationTier, format = "spreadsheet")
}
\arguments{
\item{pt}{DurationTier object}

\item{fileNameDurationTier}{file name to be created}

\item{format}{Output file format (\code{"short"} (short text format), \code{"text"} (a.k.a. full text format), \code{"spreadsheet"} (default), \code{"headerless"} (not recommended, it does not contain \code{tmin} and \code{tmax} info))}
}
\description{
Saves DurationTier to a file (in UTF-8 encoding).
\code{pt} is a list with \code{$t} and \code{$f} vectors (of the same length) at least.
If there are no \code{$tmin} and \code{$tmax} values, there are
set as min and max of \code{$t} vector.
}
\details{
Note: All credit for these functions go to Bořil, T., & Skarnitzl, R. (2016)
I needed to read in duration tiers, which have the same exact format as
pitch tiers, so all I really needed was to find-and-replace "PitchTier" with
"DurationTier" from here:
https://github.com/bbTomas/rPraat/blob/master/R/pt.R
}
\examples{
\dontrun{
pt <- dt.sample()
pt <- dt.Hz2ST(pt)    #  conversion of Hz to Semitones, reference 0 ST = 100 Hz.
dt.plot(pt)
dt.write(pt, "demo/H_st.DurationTier")
}
}
