% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erb_difference.R
\name{erb_difference}
\alias{erb_difference}
\title{Difference between frequencies in ERBs}
\usage{
erb_difference(f1, f2)
}
\arguments{
\item{f1}{Frequency 1 in Hz}

\item{f2}{Frequency 2 in Hz}
}
\value{
Difference between \code{f1} and \code{f2} on the ERB scale
}
\description{
Difference between frequencies in ERBs
}
