% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_to_erb.R
\name{erb_to_hz}
\alias{erb_to_hz}
\title{Convert ERB values to Hz values}
\usage{
erb_to_hz(erb)
}
\arguments{
\item{erb}{Numeric vector of ERB values}
}
\value{
Numeric vector of corresponding Hz values
}
\description{
Convert ERB values to Hz values
}
