% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_IntensityTier.R
\name{format.IntensityTier}
\alias{format.IntensityTier}
\title{Format console intensitytier plot}
\usage{
\method{format}{IntensityTier}(x, ..., .horiz_res = 80L, .vert_res = 10L)
}
\arguments{
\item{x}{IntensityTier from \code{it.read}}

\item{...}{Unused}

\item{.horiz_res}{Horizontal resolution}

\item{.vert_res}{Vertical resolution}
}
\value{
Returns formatted string to pass to \code{print.IntensityTier}
}
\description{
See \code{format.PitchTier}
}
