% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PitchTier.R
\name{format.PitchTier}
\alias{format.PitchTier}
\title{Format console pitchtier plot}
\usage{
\method{format}{PitchTier}(x, ..., .horiz_res = 80L, .vert_res = 10L)
}
\arguments{
\item{x}{PitchTier object}

\item{...}{Not used}

\item{.horiz_res}{Horizontal resolution, integer number of characters
wide to print. Will be coerced to be 10 if given too small a value.}

\item{.vert_res}{Vertical resolution, integer number of characters high to
print. Will be coerced to an even number if given an odd number & coerced
to 4 if given too small a value}
}
\value{
Returns formatted string to pass to \code{print.PitchTier}
}
\description{
Formats a pitchtier object's pitch track via ASCII art.
Header: Filename of the object and total number of pitch pulses
Y-Axis: Maximum and Minimum F0 values, rounded to nearest integer
X-Axis: File start time (usually 0) and end time
}
\details{
This looks like a lot but the execution time on average is only about
a millisecond.
}
