% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_TextGrid.R
\name{format.TextGrid}
\alias{format.TextGrid}
\title{Format console pitchtier plot}
\usage{
\method{format}{TextGrid}(x, ..., .horiz_res = 80L, .vert_res = 5L)
}
\arguments{
\item{x}{PitchTier object}

\item{...}{Not used}

\item{.horiz_res}{Horizontal resolution, integer number of characters
wide to print. Default is 80, recommended to be at least 50. Coerced to be 2
if given a value less than 2, although 2 isn't particularly useful}

\item{.vert_res}{Vertical resolution, integer number of characters high to
print. Will be coerced to an even number if given an odd number & coerced
to 4 if given too small a value}
}
\value{
Returns formatted string to pass to \code{print.PitchTier}
}
\description{
Formats a textgrid object to print a tiered representation to the console.
Filename is shown at the top, with tiers printed in descending order as they
would be displayed in Praat.
Interval tiers are shown with \verb{[square brackets]} while point tiers are shown
with \verb{|bars on the ends|}
}
\details{
Summary information about the number of labeled annotations out of the total
number of annotations on the tier is provided, along with the total duration.
}
