% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuclear_textgrids.R
\name{get_nuclear_textgrids}
\alias{get_nuclear_textgrids}
\title{Get nuclear textgrid region}
\usage{
get_nuclear_textgrids(textgrid_df, nuclear_words, .first = TRUE)
}
\arguments{
\item{textgrid_df}{Output of \code{batch_process_textgrids}}

\item{nuclear_words}{Character vector of nuclear words to use}

\item{.first}{Whether to only include the first result, default TRUE}
}
\value{
Filtered dataframe with the nuclear region information for each file
}
\description{
Given a dataframe of textgrid information from \code{batch_process_textgrids},
filter to only include the region corresponding to the nuclear words
specified by \code{nuclear_words}. This implicitly assumes that each target
utterance has a unique nuclear word.
}
