% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_at_time.R
\name{get_value_at_time}
\alias{get_value_at_time}
\title{Interpolate value at time}
\usage{
get_value_at_time(tier_obj, t)
}
\arguments{
\item{tier_obj}{DurationTier or PitchTier object, not to be confused with
a single tier from a textgrid}

\item{t}{time to calculate at}
}
\value{
Numeric value (pitch or duration multiplier) at time t
}
\description{
Given a pitch tier or duration tier and a timepoint, calculate the value
at that time.
}
\details{
Note to self: this is a different approach to \code{interpolate_pitchpoints},
they can and should be unified, with interpolate_pitchpoints being a
vectorized wrapped over this.
}
