% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_interval_indices_methods.R
\name{guess_interval_indices}
\alias{guess_interval_indices}
\title{Title}
\usage{
guess_interval_indices(section_column)
}
\arguments{
\item{section_column}{Numeric, Integer, or Character vector of values}
}
\value{
Integer vector of indices for each contiguous group of values
}
\description{
Title
}
