% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{guess_interval_indices_double}
\alias{guess_interval_indices_double}
\title{Double section indices}
\usage{
guess_interval_indices_double(section_column)
}
\arguments{
\item{section_column}{Vector of values}
}
\value{
Integer vector of indices for each contiguous group of values
}
\description{
Make section indices
}
\details{
Given a vector, assign contiguous groups based on where the values change.
}
