% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_to_erb.R
\name{hz_to_erb}
\alias{hz_to_erb}
\title{Convert Hz values to ERBs}
\usage{
hz_to_erb(hz_val)
}
\arguments{
\item{hz_val}{Numeric vector of Hz values}
}
\value{
Numeric vector of corresponding ERB values
}
\description{
Convert Hz values to ERBs
}
