% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_pitchpoints.R
\name{interpolate_pitchpoints}
\alias{interpolate_pitchpoints}
\title{Interpolate pitch points}
\usage{
interpolate_pitchpoints(new_times, old_times, pitch_vals)
}
\arguments{
\item{new_times}{Sorted numeric vector with range within \code{old_times}}

\item{old_times}{Sorted numeric vector}

\item{pitch_vals}{Pitch values, of same length as \code{old_times}}
}
\value{
Numeric vector of interpolated pitch points to correspond to the
timepoints at \code{new_times}
}
\description{
Given a vector of numeric timepoints \code{new_times} that lie within the range
of \code{old_times} and \code{pitch_vals} located at \code{old_times}, determine which
two pitch points are adjacent to each new timepoint, then calculate the
weighted mean based on how far the time point is from each pitch point to
interpolate between the two.
}
