% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tier_list.R
\name{is_tier_list}
\alias{is_tier_list}
\title{Valid list-representation of textgrid}
\usage{
is_tier_list(tier_list)
}
\arguments{
\item{tier_list}{Named list with relevant tier information}
}
\value{
TRUE if the list can be used as a textgrid sans class attributes
}
\description{
Checks if a given list can correspond to a valid textgrid.
Useful if you're using R to create new textgrid tiers to add to an existing
textgrid.
Note that this is not \code{is_textgrid} because it does not check for the
relevant class attributes associated with textgrids loaded with \code{rPraat::tg.read}
}
