% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tiers.R
\name{nest_tiers}
\alias{nest_tiers}
\title{Nest two textgrid tiers}
\usage{
nest_tiers(
  tg_df_list,
  nesting = "words/phones",
  filter_silence = "",
  .as_singular = TRUE
)
}
\arguments{
\item{tg_df_list}{List of textgrid dataframes, the output of
\code{textgrid_to_dataframes}}

\item{nesting}{String of the form \code{"X/Y"}, read as "Y nested within X".
Defaults to \code{"words/phones"}, intended for use with montreal forced aligner
output.}

\item{filter_silence}{Character vector of silence tokens you don't want to
keep. Defaults to \code{""}, but some people may want to use \code{c("", "<sil>")}
depending on how their textgrids are set up.}

\item{.as_singular}{Whether to remove the "s" from "words" and "phones" in
column name outpouts. Eg \code{word_start} instead of \code{words_start} for the
interval's starting timestamp.}
}
\value{
Merged dataframe
}
\description{
Given two dataframes for two different tiers, will return a single
dataframe with the dataframes merged together. This assumes that the
second tier/dataframe is nested within the first, for example a phones
tier nested within a words tier.
}
