% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_interpolate_pulses.R
\name{piecewise_interpolate_pulses}
\alias{piecewise_interpolate_pulses}
\title{Piecewise interpolation of equally spaced pitch pulses}
\usage{
piecewise_interpolate_pulses(
  pitchtier_df,
  section_by = "is_nuclear",
  pulses_per_section,
  time_by = "timepoint_norm",
  .grouping = "file",
  .pitchval = "hz",
  parallelize = FALSE
)
}
\arguments{
\item{pitchtier_df}{Pitchtier dataframe, output of \code{batch_process_pitchtiers}
and preferably with time normalization already applied}

\item{section_by}{Column name containing the section designations of the pitch contour}

\item{pulses_per_section}{An integer vector of how many points to use
for each section of the contour. This needs to have EITHER 1 value, which
is recycled for all sections OR as many values as there are sections. In the
latter case, it is HIGHLY RECOMMENDED that this vector is NAMED and in the
desired order of the sections. If the vector is unnamed, it will use the
names of the section in the order they appear in the data.}

\item{time_by}{Quoted column name containing the timepoints, defaults to
\code{"timepoint_norm"} (highly recommended to use \code{time_normalize} as a
preprocessing step beforehand). Note that the results will have new
time-normalized values with the same column name.}

\item{.grouping}{Quoted column name indexing the unique recordings, defaults
to \code{"file"}}

\item{.pitchval}{Quoted column name for which pitch values to use, defaults
to \code{"hz"}}

\item{parallelize}{Whether to run in parallel via multisession
\code{furrr::future_map_dfr}}
}
\value{
Dataframe of interpolated pitch pulses by section
}
\description{
This is an extension of \code{interpolate_equal_pulses} which extracts equal pulses
within particular sections. This was written as a component of
\code{average_pitchtracks}, where I needed to extract equally spaced pulses
within prenuclear and nuclear regions separately so I could then take
averages within the sections. This function can also be useful if you
want a lower resolution for one section but a higher resolution for another.
}
