% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchtier_to_dataframe.R
\name{pitchtier_to_dataframe}
\alias{pitchtier_to_dataframe}
\title{Convert pitchtier object to dataframe}
\usage{
pitchtier_to_dataframe(pitchtier, add_semitones = TRUE, add_erbs = TRUE, ...)
}
\arguments{
\item{pitchtier}{Pitchtier object from \code{pt.read}}

\item{add_semitones}{Logical, default TRUE, whether to add semitone
differences from a baseline}

\item{add_erbs}{Logical, default TRUE, whether to add ERB units}

\item{...}{Additional arguments passed to \code{hz_to_semitones}}
}
\value{
Dataframe representation of the given pitchtier object
}
\description{
Convert pitchtier loaded by \code{rPraat::pt.read} to a dataframe.
If semitones are included via \code{add_semitones}, a \code{semitones_from} attribute
is added. Access with \code{attr(df, 'semitones_from')}.
}
