% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pitchtracks.R
\name{preprocess_pitchtracks}
\alias{preprocess_pitchtracks}
\title{Common pitch track preprocessing steps}
\usage{
preprocess_pitchtracks(
  pitchtier_df,
  nuclear_df,
  runmed_k,
  time_normalize,
  n_pulses,
  .hz = "hz",
  .timepoint = "timepoint"
)
}
\arguments{
\item{pitchtier_df}{Pitchtier dataframe, output of \code{batch_process_pitchtiers}}

\item{nuclear_df}{Pass the output of \code{get_nuclear_tg} to code whether a pitch
pulse lies in the nuclear region or not via \code{code_nuclear_region}}

\item{runmed_k}{\code{k} for \code{runmed}, to call \code{running_median_smooth}. Changes
the Hz values.}

\item{time_normalize}{Logical, whether to normalize the pitch track to be
between (0, 1] via \code{time_normalize}. Changes the time}

\item{n_pulses}{Number of equally spaced pulses to extract via
\code{extract_equal_pulses}}

\item{.hz}{For \code{running_median_smooth}, quoted variable name for column to
hold new Hz values. Defaults to Hz.}

\item{.timepoint}{For \code{time_normalize}, quoted variable name for column to
hold new timepoint values. Defaults to timepoint.}
}
\value{
Dataframe with modified columns
}
\description{
Wrapper function around other functions to apply various preprocessing
operations to a pitchtier df containing multiple files.
If you don't want to apply a particular operation, leave it blank.
Note that the order of the arguments as presented in this description is
the order of preprocesing operations. If you do them separately, I recommend
using the order shown here.
}
