% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_IntensityTier.R
\name{print.IntensityTier}
\alias{print.IntensityTier}
\title{Print IntensityTier}
\usage{
\method{print}{IntensityTier}(x, ...)
}
\arguments{
\item{x}{IntensityTier object}

\item{...}{Not used}
}
\value{
invisibly return \code{x}
}
\description{
See \code{print.PitchTier}, this does the same thing but with intensity tiers
read in by \code{rPraat::it.read()}
}
