% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PitchTier.R
\name{print.PitchTier}
\alias{print.PitchTier}
\title{Print pitchtier object}
\usage{
\method{print}{PitchTier}(x, ...)
}
\arguments{
\item{x}{Pitchtier from \code{pt.read}}

\item{...}{Unused}
}
\value{
invisibly returns \code{x}
}
\description{
See \code{format.PitchTier}
}
