% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_median_smooth.R
\name{running_median_smooth}
\alias{running_median_smooth}
\title{Running median smooth pitch track}
\usage{
running_median_smooth(pitchtier_df, .k = 5, .hz = "hz")
}
\arguments{
\item{pitchtier_df}{Pitchtier dataframe, output of \code{batch_process_pitchtiers}}

\item{.k}{\code{k} to use for \code{runmed}}

\item{.hz}{Quoted variable name to hold new Hz values. Defaults to \code{hz},
which will override the previous values}
}
\value{
\code{pitchtier_df} with modified \code{hz} column or new column specified by
\code{.hz}
}
\description{
Given a pitchtier dataframe, smooth the Hz values using a running median
}
