% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_interval_tier_by_dt.R
\name{scale_interval_tier_by_dt}
\alias{scale_interval_tier_by_dt}
\title{Scale interval tier by DurationTier}
\usage{
scale_interval_tier_by_dt(int_tier, durationtier)
}
\arguments{
\item{int_tier}{Interval tier, i.e., \code{tg[[x]]}}

\item{durationtier}{DurationTier object from \code{dt.read}}
}
\value{
Interval tier with new t1 and t2 values
}
\description{
Part of \code{scale_textgrid_by_dt}, this is the procedure for a singular
interval tier
}
