% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_point_tier_by_dt.R
\name{scale_point_tier_by_dt}
\alias{scale_point_tier_by_dt}
\title{Scale point tier by DurationTier}
\usage{
scale_point_tier_by_dt(point_tier, durationtier)
}
\arguments{
\item{point_tier}{Singular point tier from a textgrid, i.e. \code{tier[[x]]}}

\item{durationtier}{DurationTier object from \code{dt.read}}
}
\value{
The same point tier but with shifted times
}
\description{
Part of \code{scale_textgrid_by_dt}, this is the procedure for a singular
point tier
}
