% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_textgrid_by_dt.R
\name{scale_textgrid_by_dt}
\alias{scale_textgrid_by_dt}
\title{Scale textgrid by duration tier}
\usage{
scale_textgrid_by_dt(textgrid, durationtier)
}
\arguments{
\item{textgrid}{TextGrid object from \code{tg.read}}

\item{durationtier}{DurationTier object from \code{dt.read}}
}
\value{
The textgrid with modified time values for each tier
}
\description{
Given a textgrid and a duration tier, the interval boundaries and point
locations will be shifted over in accordance with sections that shrink
and stretch. This is most useful when you know exactly how you want
to manipulate the duration of a sound file but want to maintain any
textgrid that goes along with it in the warped file.
}
