% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_fileinfo.R
\name{separate_fileinfo}
\alias{separate_fileinfo}
\title{Separate fileinfo}
\usage{
separate_fileinfo(
  dat,
  .col = "file",
  .sep = "_",
  .into = c("utterance", "tune", "take")
)
}
\arguments{
\item{dat}{Textgrid or pitchtier dataframe}

\item{.col}{See \code{?tidyr::separate}}

\item{.sep}{See \code{?tidyr::separate}}

\item{.into}{See \code{?tidyr::separate}}
}
\value{
\code{dat} with columns specified by .into
}
\description{
Wrapper around tidyr::separate
}
