% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_normalize.R
\name{time_normalize}
\alias{time_normalize}
\title{Time normalize pitch track}
\usage{
time_normalize(pitchtier_df, .timepoint = "timepoint")
}
\arguments{
\item{pitchtier_df}{Pitchtier dataframe, output of \code{batch_process_pitchtiers}}

\item{.timepoint}{Quoted variable name to hold new timepoint values.
Defaults to \code{timepoint}, which will override the previous values.}
}
\value{
\code{pitchtier_df} with modified \code{timepoint} column or new column
specified by \code{.timepoint}
}
\description{
Changes the timepoints to be between (0, 1]
}
