% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_praat_arguments.R
\name{validate_praat_arguments}
\alias{validate_praat_arguments}
\title{Put praatscript arguments in correct order}
\usage{
validate_praat_arguments(script_path, use_defaults, ...)
}
\arguments{
\item{script_path}{Path to the script}

\item{use_defaults}{Logical, whether to fill with default values or not.}

\item{...}{\emph{Named} list of arguments to pass to the script. If any character
arguments contain spaces, they will be escaped for you.}
}
\value{
Character vector of values to be passed to a command line
praatscript invocation via Praat --run
}
\description{
Given a script and named arguments, slot the arguments into the correct
position based on where they appear in the script. See \code{run_praatscript} for
more information. Typically you won't need to use this yourself, as it's a
helper for \code{run_praatscript}.
}
