% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_textgrid.R
\name{write_textgrid}
\alias{write_textgrid}
\title{Write textgrids to directory}
\usage{
write_textgrid(textgrid, directory, filename = NULL, format = "short")
}
\arguments{
\item{textgrid}{TextGrid object}

\item{directory}{Directory to write textgrid file to}

\item{filename}{New filename for the textgrid file. If \code{NULL}, the default,
the file name will be extracted from the textgrid object.}

\item{format}{See \code{rPraat::tg.write}, defaults to "short"}
}
\value{
Invisibly returns the new file path
}
\description{
A  wrapper for \code{rPraat::tg.write} that will write a textgrid to the given
directory with the filename pulled from the class attributes.
Note that the filepath will be quietly sanitized with \code{fs::path_sanitize}
to ensure that the filename is valid.
}
