% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{wrLine}
\alias{wrLine}
\title{wrLine}
\usage{
wrLine(string, fid, collectionFullText = FALSE)
}
\arguments{
\item{string}{Text line.}

\item{fid}{A connection object.}

\item{collectionFullText}{If TRUE, add extra 8 spaces at the beginning of the string}
}
\value{
a raw vector (if fid is a raw vector) or invisibly NULL.
}
\description{
Taken from: https://github.com/bbTomas/rPraat/blob/master/R/rpraat.R
because it wasn't exported from the package
}
\details{
Write text line to a connection in binary mode.
}
