% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_pitchtracks.R
\name{average_pitchtracks}
\alias{average_pitchtracks}
\title{Equally-spaced average pitch contours}
\usage{
average_pitchtracks(
  raw_pitchtier_df,
  section_by = "is_nuclear",
  pulses_per_section = c(`FALSE` = 15, `TRUE` = 30),
  time_by = "timepoint_norm",
  aggregate_by = take ~ utterance + tune,
  .pitchval = "hz"
)
}
\arguments{
\item{raw_pitchtier_df}{Pitchtier dataframe, output of \code{batch_process_pitchtiers}
and preferably with time normalization already applied}

\item{section_by}{Column name containing the section designations of the pitch contour}

\item{pulses_per_section}{An integer vector of how many points to use
for each section of the contour. This needs to have EITHER 1 value, which
is recycled for all sections OR as many values as there are sections. In the
latter case, it is HIGHLY RECOMMENDED that this vector is NAMED and in the
desired order of the sections. If the vector is unnamed, it will use the
names of the section in the order they appear in the data.}

\item{time_by}{Quoted column name containing the timepoints, defaults to
\code{"timepoint_norm"} (highly recommended to use \code{time_normalize} as a
preprocessing step beforehand)}

\item{aggregate_by}{A formula of unquoted column names. The LHS should index
individual recordings, e.g. a \code{file} column. The RHS should index what you
want to aggregate over, e.g. by \code{utterance} and \code{tune}. This would be written
as \code{file ~ utterance + tune}. Unpredictable results may occur if recordings
are not uniquely identifiable or if the LHS is not specified.}

\item{.pitchval}{Quoted column name containing the pitch values to average
over. Defaults to \code{"hz"}}
}
\value{
A dataframe containing the averaged and equally spaced piecewise
pitch contours
}
\description{
This function averages pitch contours, with the possibility of doing the
averaging by section. For example, we could average the prenuclear portion
with 15 equally spaced points and then average the nuclear portion with 30
equally spaced points. This avoids the problem of setting 45 points for the
total utterance but being unsure about whether nuclear points are being
averaged with prenuclear points.
}
\details{
You can also chain this for higher-order aggregation, e.g. to get the average
tune produced by a participant, then the average across all participants:
pitchtier_df |>
average_pitchtracks(aggregate_by = file ~ utterance + tune + participant) |>
average_pitchtracks(aggregate_by = participant ~ utterance + tune)
}
