% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_pitchtiers.R
\name{batch_process_pitchtiers}
\alias{batch_process_pitchtiers}
\title{Load multiple pitchtiers into wrangled dataframe}
\usage{
batch_process_pitchtiers(ptdir)
}
\arguments{
\item{ptdir}{Directory containing pitchtier files}
}
\value{
Dataframe of the pitchtier information from each file
}
\description{
Wrapper around filepath |> pitchtier_to_dataframe()  |> separate_fileinfo(),
merged together via \code{purrr::map_dfr}
}
