% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_textgrids.R
\name{batch_process_textgrids}
\alias{batch_process_textgrids}
\title{Load multiple textgrids into wrangled dataframe}
\usage{
batch_process_textgrids(tgdir)
}
\arguments{
\item{tgdir}{Directory containing textgrids}
}
\value{
Dataframe of the word and phone information from each textgrid
}
\description{
Wrapper around filepath |> textgrid_to_dataframes() |> nest_tiers() |>
separate_fileinfo(), merged together via \code{purrr::map_dfr}
}
\details{
Note: The textgrids are assumed (for now) to only contain 2 interval tiers:
\code{words} and \code{phones}, as given by the output of the Montreal forced aligner.
}
