% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_TextGrid.R
\name{print.TextGrid}
\alias{print.TextGrid}
\title{Print pitchtier object}
\usage{
\method{print}{TextGrid}(x, ...)
}
\arguments{
\item{x}{Pitchtier from \code{pt.read}}

\item{...}{Unused}
}
\value{
invisibly returns \code{x}
}
\description{
See \code{format.PitchTier}
}
