% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_median_smooth.R
\name{running_median_smooth}
\alias{running_median_smooth}
\title{Running median smooth pitch track}
\usage{
running_median_smooth(
  pitchtier_df,
  .k = 5,
  .from = "hz",
  .to = "runmed_hz",
  .grouping = "file"
)
}
\arguments{
\item{pitchtier_df}{Pitchtier dataframe, output of \code{batch_process_pitchtiers}}

\item{.k}{\code{k} to use for \code{runmed}}

\item{.from}{Quoted variable name to pull Hz values from. Defaults to \code{.hz}}

\item{.to}{Quoted variable name to put smoothed Hz values. Defaults to
\code{runmed_hz}. If this is the same as \code{.from}, the values will be overwritten.}

\item{.grouping}{Character vector of columns to group by, defaults to \code{'file'}}
}
\value{
\code{pitchtier_df} with modified \code{hz} column or new column specified by
\code{.hz}
}
\description{
Given a pitchtier dataframe, smooth the Hz values using a running median
}
